Changed bullet to be alive for 0.4s instead of 1s

+++ Assets/Scripts/Enemies/NeedleShooter/NeedleShooterShot.cs
@@ -9,7 +9,7 @@ public class NeedleShooterShot : MonoBehaviour
   /// Damage inflicted
   /// </summary>
   public int damage = 1;
-    public int bulletTimeAlive = 2; // time bullet stays alive in seconds
+	public float bulletTimeAlive = 2f; // time bullet stays alive in seconds
 
 
     /// <summary>
